requirements.txt
python-telegram-bot==21.4
feedparser==6.0.11
tzdata==2025.2

